move INFFAST_CHUCKSIZE code to memcopy.h
